refactor : entity 패키지 상위레벨로 변경
