+++ src/main/java/subway/application/StationService.java
@@ -8,9 +8,9 @@ import subway.domain.Station;
 import subway.dto.SectionRequest;
 import subway.dto.StationRequest;
 import subway.dto.StationResponse;
-import subway.persistence.entity.LineEntity;
-import subway.persistence.entity.SectionEntity;
-import subway.persistence.entity.StationEntity;
+import subway.entity.LineEntity;
+import subway.entity.SectionEntity;
+import subway.entity.StationEntity;
 import subway.persistence.repository.StationRepository;
 
 import java.util.List;
