+++ src/main/java/subway/entity/SectionEntity.java
@@ -1,4 +1,4 @@
-package subway.persistence.entity;
+package subway.entity;
 
 public class SectionEntity {
     private Long id;
