+++ src/main/java/subway/entity/StationEntity.java
@@ -1,4 +1,4 @@
-package subway.persistence.entity;
+package subway.entity;
 
 public class StationEntity {
     private Long id;
