+++ src/main/java/subway/persistence/dao/StationDao.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.persistence.entity.StationEntity;
+import subway.entity.StationEntity;
 
 import java.util.List;
