+++ src/main/java/subway/persistence/repository/JdbcLineRepository.java
@@ -1,12 +1,12 @@
 package subway.persistence.repository;
 
 import org.springframework.stereotype.Repository;
+import subway.entity.LineEntity;
+import subway.entity.SectionEntity;
+import subway.entity.StationEntity;
 import subway.persistence.dao.LineDao;
 import subway.persistence.dao.SectionDao;
 import subway.persistence.dao.StationDao;
-import subway.persistence.entity.LineEntity;
-import subway.persistence.entity.SectionEntity;
-import subway.persistence.entity.StationEntity;
 
 import java.util.List;
